+++ inc-setup/admin-bar-items.php
@@ -114,12 +114,9 @@ function _mw_adminimize_change_admin_bar() {
 		}
 	}
 
-	// New declaration of type for array reduce.
-	if ( ! isset( $disabled_admin_bar_option_ ) ) {
-		$disabled_admin_bar_option_ = array();
-	}
+	var_dump($disabled_admin_bar_option_);
 	// Merge multidimensional array in to one, flat.
-	$disabled_admin_bar_option_ = (array) array_reduce( $disabled_admin_bar_option_, 'array_merge', array() );
+	$disabled_admin_bar_option_ = (array) array_reduce( (array) $disabled_admin_bar_option_, 'array_merge', array() );
 
 	// Support Multiple Roles for users.
 	if ( _mw_adminimize_get_option_value( 'mw_adminimize_multiple_roles' ) && 1 < count( $user->roles ) ) {