+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file. This projec
 * Fix open Translations. props pedro-mendonca
 * Fix Typos.
 * Fix php warning on Admin Bar items.
+* Fix CPT feature support, if it false.
 
 ## [1.10.6](https://github.com/bueltge/Adminimize/compare/1.10.5...1.10.6) - 2016-08-09
 ### Fixed