+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file. This projec
 ## [Unreleased](https://github.com/bueltge/Adminimize/compare/1.10.6...HEAD)
 * Fix open Translations. props pedro-mendonca
 * Fix Typos.
-* Fix php warning on Admin Bar items.
+* Fix php warning on Admin Bar items for PHP 5.2.
 * Fix CPT feature support, if it false.
+* Add check in different functions for AJAX request.
 
 ## [1.10.6](https://github.com/bueltge/Adminimize/compare/1.10.5...1.10.6) - 2016-08-09
 ### Fixed